Adjusting how TokenType is setup
TokenType is now setup as a string alias and the different tokens are setup as individual const variables instead of the enum. This way we can use the tokens in the toString function to print out their name and other future uses.

We also properly setup the const variables (type, lexeme, literal, line) in the constructor.